Move enrollment barriers into enrollment library
