+++ apps/prairielearn/src/lib/enrollment/barrier.sql
